change motulator for gritulator in directories names
