+++ gritulator/_helpers.py
@@ -22,7 +22,7 @@ def abc2complex(u):
 
     Examples
     --------
-    >>> from motulator import abc2complex
+    >>> from gritulator import abc2complex
     >>> y = abc2complex([1, 2, 3])
     >>> y
     (-1-0.5773502691896258j)
@@ -48,7 +48,7 @@ def complex2abc(u):
 
     Examples
     --------
-    >>> from motulator import complex2abc
+    >>> from gritulator import complex2abc
     >>> y = complex2abc(1-.5j)
     >>> y
     array([ 1.       , -0.9330127, -0.0669873])
