+++ gritulator/control/__init__.py
@@ -1,10 +1,10 @@
 """This package contains example controllers."""
 
-from motulator.control._common import (
+from gritulator.control._common import (
     ComplexPICtrl, RateLimiter, SpeedCtrl, PICtrl, PWM)
 
-import motulator.control.gfl as gfl
-import motulator.control.gfm as gfm
+import gritulator.control.gfl as gfl
+import gritulator.control.gfm as gfm
 
 __all__ = [
     "ComplexPICtrl", "RateLimiter", "SpeedCtrl", "PICtrl", "PWM", "gfl", "gfm"
