+++ gritulator/control/_common.py
@@ -1,8 +1,8 @@
 """Common control functions and classes."""
 
 import numpy as np
-from motulator._helpers import abc2complex, complex2abc
-from motulator._utils import Bunch
+from gritulator._helpers import abc2complex, complex2abc
+from gritulator._utils import Bunch
 
 
 # %%
