+++ gritulator/control/gfl/__init__.py
@@ -1,8 +1,8 @@
 """This package contains example controllers for grid following converters."""
 
-from motulator.control.gfl._grid_following import (
+from gritulator.control.gfl._grid_following import (
     GridFollowingCtrl,
     GridFollowingCtrlPars,
 )
 
-# from motulator.control._common import RateLimiter, SpeedCtrl
+# from gritulator.control._common import RateLimiter, SpeedCtrl
