+++ gritulator/control/gfm/_power_synchronization.py
@@ -15,9 +15,9 @@ from __future__ import annotations
 from collections.abc import Callable
 from dataclasses import dataclass, field
 import numpy as np
-from motulator._helpers import abc2complex
-from motulator._utils import Bunch
-from motulator.control._common import Ctrl, PWM
+from gritulator._helpers import abc2complex
+from gritulator._utils import Bunch
+from gritulator.control._common import Ctrl, PWM
 
 # %%
 @dataclass
