+++ gritulator/model/_grid_filter.py
@@ -7,7 +7,7 @@ stationary coordinates.
 
 """
 import numpy as np
-from motulator._helpers import complex2abc
+from gritulator._helpers import complex2abc
 
 
 # %%
