+++ gritulator/model/dc_bus/__init__.py
@@ -1,8 +1,8 @@
 """Continuous-time DC-bus models."""
 
-from motulator.model.dc_bus._dc_bus import DCBus
+from gritulator.model.dc_bus._dc_bus import DCBus
 
-from motulator.model.dc_bus._dc_dyn_model import (
+from gritulator.model.dc_bus._dc_dyn_model import (
     DCBusAndLFilterModel,
     DCBusAndLCLFilterModel,
 )
