--- motulator/__init__.py
@@ -1,28 +1,0 @@
-"""
-*motulator*: Motor Drive Simulator in Python
-
-This software includes continuous-time simulation models for induction machines 
-and synchronous machines. Furthermore, selected examples of discrete-time 
-control algortihms are also included as well as various utilities.
-
-"""
-
-from motulator._helpers import (
-    abc2complex, complex2abc, BaseValues, BaseValuesElectrical, Sequence, Step)
-
-from motulator._plots import plot_grid
-
-from motulator import control
-from motulator import model
-
-__all__ = [
-    "abc2complex",
-    "complex2abc",
-    "BaseValues",
-    "BaseValuesElectrical",
-    "Sequence",
-    "Step",
-    "control",
-    "model",
-    "plot_grid",
-]
